+++ lib/adxo.py
@@ -16,6 +16,7 @@ def remove_control_characters(s):
 
 def get_adxo_events():
     # URL del file XML RSS
+
     rss_url = "https://www.ng3k.com/adxo.xml"    
 
     try:
@@ -34,8 +35,20 @@ def get_adxo_events():
         for item in feed.entries:
             prop = {}
             title = item.title
-            #title = "Sint Maarten: Dec 2 2023 - Jan 20 2024 -- PJ7AA -- QSL via: LoTW            "
+
+            #bash debug command for list how many commas for each title
+            #curl https://www.ng3k.com/adxo.xml | grep title | awk -F',' '{print "Row " NR " has " NF-1 " comma: " $0}'      
+            
+            #if there are two commas replace them with space in order to normalize a string like this:
+            #       "Dominica: Dec 26, 2024-Jan 4, 2025 -- J75K -- QSL via: LoTW       "
+            #it will be
+            #       "Dominica: Dec 26  2024-Jan 4  2025 -- J75K -- QSL via: LoTW       "
+
+            logging.debug('===========================================================')
             logging.debug(title)
+
+            if title.count(',') == 2: 
+                title=title.replace(',', ' ')      
 
             #callsign
             start_callsign_idx = title.find("--")
@@ -44,14 +57,17 @@ def get_adxo_events():
            
             #period
             period = title[title.find(":")+1: start_callsign_idx] 
-            comma_year_idx = period.find(",")
 
+            comma_year_idx = period.find(",")
             #start date - end date
             if comma_year_idx > 0:
                 #Mar 23-Apr 1, 2024 or Mar 23-30, 2024 
                 year = period[comma_year_idx+1:].strip()
+                logging.debug("year: {}".format(year))
                 date_start = period[:period.find("-")]+" "+year
+                logging.debug("date_start: {}".format(date_start))
                 date_end = period[period.find("-")+1:comma_year_idx]+" "+year
+                logging.debug("date_end: {}".format(date_end))                
                 match = re.search(r"^([A-Za-z]{3}) \d{1,2} \d{4}$", date_end)
                 if match:
                     #Mar 23-Apr 1, 2024    
@@ -59,14 +75,27 @@ def get_adxo_events():
                 else:
                     #Mar 23-30, 2024 
                     date_end=date_start[:5]+date_end                 
+                    logging.debug("date_end: {}".format(date_end))                
             else:
                 #Mar 23 2023-Apr 1 2024 
                 date_start = period[:period.find("-")]
+                logging.debug("date_start: {}".format(date_start))
                 date_end = period[period.find("-")+1:]
+                logging.debug("date_end: {}".format(date_end))          
 
-            prop["start"]  = datetime.strptime(date_start.strip(), "%b %d %Y")
-            prop["end"] = datetime.strptime(date_end.strip(), "%b %d %Y")
-            
+            try:
+                prop["start"]  = datetime.strptime(date_start.strip(), "%b %d %Y")
+                prop["end"] = datetime.strptime(date_end.strip(), "%b %d %Y")
+
+            except Exception as eDate:
+                logging.error("Dates not decoded:")
+                logging.error(eDate)
+                logging.error("Date start: {}".format(date_start))
+                logging.error("Date end: {}".format(date_end))
+                #since is not decoded assign current date
+                prop["start"]= datetime.now()
+                prop["end"]= datetime.now()
+
             prop["summary"] = remove_control_characters(title)
             prop["description"] = remove_control_characters(item.description)
 
@@ -76,12 +105,15 @@ def get_adxo_events():
             #append only valids (in date) events
             if prop["start"] <= now  and prop["end"] >= now:
                 events.append(prop)
+                logging.debug("Event added")
+            else:
+                logging.debug("Discarted because not in current date range")
 
         logging.debug(events)
         if len(events) > 0:
             logging.info("number ADXO events: " + str(len(events)))
         else:
-            logging.warn("No ADXO events founds")
+            logging.warning("No ADXO events founds")
 
         return events